+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## 6.8.0 (2023-06-12)
+
+### New features
+
+The result of `Completion.info` may now include a `destroy` method that will be called when the tooltip is removed.
+
+## 6.7.1 (2023-05-13)
+
+### Bug fixes
+
+Fix a bug that cause incorrect ordering of completions when some results covered input text and others didn't.
+
+## 6.7.0 (2023-05-11)
+
+### New features
+
+The new `hasNextSnippetField` and `hasPrevSnippetField` functions can be used to figure out if the snippet-field-motion commands apply to a given state.
+
 ## 6.6.1 (2023-05-03)
 
 ### Bug fixes